+++ src/main/java/lotto/Inputer.java
@@ -1,21 +1,38 @@
 package lotto;
 
+import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
+
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Scanner;
 
 public class Inputer {
 
-    public int inputMoney(){
 
+    public int inputMoney() {
 
-        return 0;
+        Scanner scanner = new Scanner(System.in);
+
+        int userInputMoney = scanner.nextInt();
+
+        return userInputMoney;
+
     }
 
-    public List<Integer> inputLottoWinNumber (){
-        return null;
+    public String createLottoWinNumber() {
+        Scanner scanner = new Scanner(System.in);
+
+        String lottoWin = scanner.next();
+
+        return lottoWin;
     }
 
-    public int inputLottoBonusNumber(){
-        return 0;
+    public int inputLottoBonusNumber() {
+        Scanner scanner = new Scanner(System.in);
+
+        int bonusNumber = scanner.nextInt();
+
+        return bonusNumber;
     }
 
 }