+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+#### 1.0.0 (2022-11-13)
+
+##### Documentation Changes
+
+* **CHANGELOG:**  change change log.md (0f95dd01)
+* **README:**
+  *  add temporary class and method name (5ef4e299)
+  *  write implementation list (abf6886c)
+
+##### New Features
+
+* **Printer:**
+  *  print each win statement and win number (66bb9421)
+  *  print revenue percent (b4c967ca)
+  *  print lotto list by ascending order (47cc7202)
+* **Calculater:**  calculate revenue (92ebc2ef)
+* **CompareHoewManyCorrect:**  change duplication number list for indexing (2387982b)
+* **CompareHowManyCorrect:**  check bonus number is in user lotto number (5e784f63)
+* **Comparator:**  compare how many correct number exists and add it list (b8455702)
+* **Generator:**
+  *  change lotto number list generator to 2 dimension list (a609585c)
+  *  implement random lotto list number generation (6 numbers) (6f4e208a)
+* **Maker:**  implement string to list method (b0d0b8ad)
+* **Inputer:**
+  *  implement input user money, win number, bonus number (8773b043)
+  *  implement user money input method (31eac496)
+*  generate class and method framework (63e975fb)
+*  setup precourse lotto project (8e048b15)
+
+##### Refactors
+
+* **CompareHowManyCorrect:**
+  *  split 'gather two list(user's number, win number) method' (2bb2fecf)
+  *  split 'count duplication method' (aeefc8ad)
+
+##### Code Style Changes
+
+*  change class name (5ba64d08)
+*  rename and restruct (c3503db7)
+
+##### Tests
+
+* **LottoTest:**
+  *  generate amount each win list (1e73d508)
+  *  test return true when certain number is in integer list (3dd53a45)
+  *  test gather two each lists (f8c3ab7b)
+  *  test count duplication (53b0c6bb)
+  *  test compareLotto method do right (0b4079a4)
+  *  test amount of generate 6 random number list account to input money (ee5124d3)
+  *  test generate 6 random number (75a7492f)
+  *  test string to list method int Maker class (eea871f7)
+
 #### 1.0.0 (2022-11-12)
 
 ##### Documentation Changes